pyproject.toml
[project]
name = "qq-music-api"
version = "0.1.0"
description = "QQ Music API implementation in Python"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
qq-music-api = "qq_music_api.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/qq_music_api"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
